track both terminals in the TMNotebookTab
Since each of the two terminals have a ScrolledWindow associated with
them, this required moving the ScrolledWindow field from the
TMNotebookTab to the TMTerm.

More importantly, this refactoring revealed many places where the two
terminals were treated differently, and that code was changed to treat
both terminals the same. This was made easier by the addition of a lens
pointing to the focused terminal (which is currently always the left
terminal), and a traversal pointing at both terminals.