Refactor TMState, adding a collection of ApplicationWindows.
This is the initial refactoring that will allow Termonad to use multiple
ApplicationWindows.

This commit just refactors the underlying TMState data types, adding a
TMWindow data type to represent a GTK ApplicationWindow.

None of the actual logic has been changed, so with just this commit,
Termonad is still only capable of using a single ApplicationWindow.